Removed item interaction code.
