+++ src/main/java/com/atherys/towns/AtherysTowns.java
@@ -255,10 +255,6 @@ public class AtherysTowns {
         return components.pollFacade;
     }
 
-    public ProtectionFacade getProtectionFacade() {
-        return components.protectionFacade;
-    }
-
     public TownsCache getTownsCache() {
         return components.townsCache;
     }
@@ -334,9 +330,6 @@ public class AtherysTowns {
 
         @Inject
         private PollFacade pollFacade;
-
-        @Inject
-        private ProtectionFacade protectionFacade;
 
         @Inject
         private PlayerListener playerListener;
