+++ src/main/java/com/atherys/towns/AtherysTownsModule.java
@@ -51,6 +51,5 @@ public class AtherysTownsModule extends AbstractModule {
         bind(TownsMessagingFacade.class);
         bind(TownAdminFacade.class);
         bind(PollFacade.class);
-        bind(ProtectionFacade.class);
     }
 }
