+++ src/main/java/com/atherys/towns/api/permission/WorldPermissionRegistryModule.java
@@ -22,8 +22,7 @@ public class WorldPermissionRegistryModule implements CatalogRegistryModule<Worl
                 WorldPermissions.INTERACT_TILE_ENTITIES,
                 WorldPermissions.INTERACT_DOORS,
                 WorldPermissions.INTERACT_REDSTONE,
-                WorldPermissions.INTERACT_ENTITIES,
-                WorldPermissions.USE_ITEMS
+                WorldPermissions.INTERACT_ENTITIES
         );
     }
